3rd iteration: Atbash Cipher @Rust
Moving common functionality of encode and decode to a helper function.
I think this improves the naming a lot. Because it might have been
confusing why encoding required to decode first. With the separation I
was able to give the common step a name, that might better explain
what's going on.